Sacado:  Properly handle teuchos sub-package dependencies.
This fixes Sacado's dependencies on Teuchos to properly only depend on
the needed Teuchos sub-packages.  This involved a fair amount of code
reorganization because several optional dependencies were combined in
one file.  There is some danger in this in that I didn't test every
possible combination of Teuchos sub-packages.  However I did test no
packages, all packages, and enabling each package individually.

This should resolve the duplicate symbol errors observed in issues #2891
and #4457 by removing the duplicate link of teuchos-core.  This was also
requested in issue #1263.
