+++ packages/sacado/cmake/Dependencies.cmake
@@ -1,7 +1,7 @@
 SET(LIB_REQUIRED_DEP_PACKAGES)
-SET(LIB_OPTIONAL_DEP_PACKAGES Teuchos KokkosCore TeuchosKokkosComm)
+SET(LIB_OPTIONAL_DEP_PACKAGES KokkosCore TeuchosCore TeuchosNumerics TeuchosComm TeuchosKokkosComm)
 SET(TEST_REQUIRED_DEP_PACKAGES)
-SET(TEST_OPTIONAL_DEP_PACKAGES TeuchosCore KokkosContainers)
+SET(TEST_OPTIONAL_DEP_PACKAGES KokkosContainers)
 SET(LIB_REQUIRED_DEP_TPLS)
 SET(LIB_OPTIONAL_DEP_TPLS)
 SET(TEST_REQUIRED_DEP_TPLS)
