+++ packages/sacado/example/CMakeLists.txt
@@ -131,7 +131,7 @@ IF (Sacado_ENABLE_CXX11)
     )
 ENDIF()
 
-IF(Sacado_ENABLE_Teuchos)
+IF(Sacado_ENABLE_TeuchosNumerics)
 
   TRIBITS_ADD_EXECUTABLE_AND_TEST(
     blas_example
@@ -149,24 +149,24 @@ IF(Sacado_ENABLE_Teuchos)
     NUM_MPI_PROCS 1
     )
 
-  IF (Sacado_ENABLE_KokkosCore)
+ENDIF()
+
+IF (Sacado_ENABLE_KokkosCore)
+
+  IF (Sacado_ENABLE_TeuchosCore)
     TRIBITS_ADD_EXECUTABLE(
       view_factory_example
       SOURCES view_factory_example.cpp
       COMM serial mpi
       )
-
-    TRIBITS_ADD_EXECUTABLE_AND_TEST(
-      dfad_view_handle_example
-      SOURCES dfad_view_handle_example.cpp
-      COMM serial mpi
-      NUM_MPI_PROCS 1
-      )
   ENDIF()
 
-ENDIF()
-
-IF (Sacado_ENABLE_KokkosCore)
+  TRIBITS_ADD_EXECUTABLE_AND_TEST(
+    dfad_view_handle_example
+    SOURCES dfad_view_handle_example.cpp
+    COMM serial mpi
+    NUM_MPI_PROCS 1
+    )
 
   TRIBITS_ADD_EXECUTABLE_AND_TEST(
     view_example
