+++ packages/sacado/src/Sacado_ELRFad_DFadTraits.hpp
@@ -138,15 +138,14 @@ namespace Sacado {
 
 } // namespace Sacado
 
+//
 // Define Teuchos traits classes
-#ifdef HAVE_SACADO_TEUCHOS
+//
+
+// Promotion traits
+#ifdef HAVE_SACADO_TEUCHOSNUMERICS
 #include "Teuchos_PromotionTraits.hpp"
-#include "Teuchos_ScalarTraits.hpp"
-#include "Sacado_Fad_ScalarTraitsImp.hpp"
-
 namespace Teuchos {
-
-  //! Specialization of %Teuchos::PromotionTraits to DFad types
   template <typename ValueT>
   struct PromotionTraits< Sacado::ELRFad::DFad<ValueT>,
                           Sacado::ELRFad::DFad<ValueT> > {
@@ -155,35 +154,42 @@ namespace Teuchos {
     promote;
   };
 
-  //! Specialization of %Teuchos::PromotionTraits to DFad types
   template <typename ValueT, typename R>
   struct PromotionTraits< Sacado::ELRFad::DFad<ValueT>, R > {
     typedef typename Sacado::Promote< Sacado::ELRFad::DFad<ValueT>, R >::type
     promote;
   };
 
-  //! Specialization of %Teuchos::PromotionTraits to DFad types
   template <typename L, typename ValueT>
   struct PromotionTraits< L, Sacado::ELRFad::DFad<ValueT> > {
   public:
     typedef typename Sacado::Promote< L, Sacado::ELRFad::DFad<ValueT> >::type
     promote;
   };
+}
+#endif
 
-  //! Specializtion of %Teuchos::ScalarTraits
+// Scalar traits
+#ifdef HAVE_SACADO_TEUCHOSCORE
+#include "Sacado_Fad_ScalarTraitsImp.hpp"
+namespace Teuchos {
   template <typename ValueT>
   struct ScalarTraits< Sacado::ELRFad::DFad<ValueT> > :
     public Sacado::Fad::ScalarTraitsImp< Sacado::ELRFad::DFad<ValueT> >
   {};
+}
+#endif
 
-  //! Specialization of %Teuchos::SerializationTraits
+// Serialization traits
+#ifdef HAVE_SACADO_TEUCHOSCOMM
+#include "Sacado_Fad_SerializationTraitsImp.hpp"
+namespace Teuchos {
   template <typename Ordinal, typename ValueT>
   struct SerializationTraits<Ordinal, Sacado::ELRFad::DFad<ValueT> > :
     public Sacado::Fad::SerializationTraitsImp< Ordinal,
                                                 Sacado::ELRFad::DFad<ValueT> >
   {};
 
-  //! Specialization of %Teuchos::ValueTypeSerializer
   template <typename Ordinal, typename ValueT>
   struct ValueTypeSerializer<Ordinal, Sacado::ELRFad::DFad<ValueT> > :
     public Sacado::Fad::SerializerImp< Ordinal,
@@ -198,6 +204,11 @@ namespace Teuchos {
       Base(vs, sz) {}
   };
 }
-#endif // HAVE_SACADO_TEUCHOS
+#endif
+
+// KokkosComm
+#if defined(HAVE_SACADO_KOKKOSCORE) && defined(HAVE_SACADO_TEUCHOSKOKKOSCOMM) && defined(HAVE_SACADO_VIEW_SPEC) && !defined(SACADO_DISABLE_FAD_VIEW_SPEC)
+#include "KokkosExp_View_Fad.hpp"
+#endif
 
 #endif // SACADO_ELRFAD_DFADTRAITS_HPP
