+++ packages/sacado/test/TestSuite/CMakeLists.txt
@@ -13,7 +13,7 @@ IF (TPL_ENABLE_ADOLC)
 
 ENDIF()
 
-IF (${PACKAGE_NAME}_ENABLE_Teuchos)
+IF (${PACKAGE_NAME}_ENABLE_TeuchosCore AND ${PACKAGE_NAME}_ENABLE_TeuchosNumerics)
 
   SET(${PACKAGE_NAME}_TESTSUITE_SOURCES ${${PACKAGE_NAME}_TESTSUITE_SOURCES}
     FadBLASUnitTests.hpp
