+++ packages/sacado/test/UnitTests/CMakeLists.txt
@@ -5,79 +5,85 @@ SET(UTILS_SOURCES
   ${PACKAGE_SOURCE_DIR}/test/utils/Sacado_RandomImp.hpp
 )
 
-# All of these tests use the Teuchos test harness, and therefore Teuchos
+# All of these tests use the Teuchos test harness, and therefore TeuchosCore
 # must be enabled
-IF (Sacado_ENABLE_Teuchos)
-  TRIBITS_ADD_EXECUTABLE_AND_TEST(
-    FadSerializationTests
-    SOURCES Fad_SerializationTests.cpp ${UTILS_SOURCES}
-    ARGS
-    COMM serial mpi
-    NUM_MPI_PROCS 1
-    STANDARD_PASS_OUTPUT
-  )
+IF (Sacado_ENABLE_TeuchosCore)
 
-  TRIBITS_ADD_EXECUTABLE_AND_TEST(
-    TaySerializationTests
-    SOURCES Tay_SerializationTests.cpp ${UTILS_SOURCES}
-    ARGS
-    COMM serial mpi
-    NUM_MPI_PROCS 1
-    STANDARD_PASS_OUTPUT
-    XHOSTTYPE Windows Windows_NT
-  )
+  IF (Sacado_ENABLE_TeuchosComm)
+    TRIBITS_ADD_EXECUTABLE_AND_TEST(
+      FadSerializationTests
+      SOURCES Fad_SerializationTests.cpp ${UTILS_SOURCES}
+      ARGS
+      COMM serial mpi
+      NUM_MPI_PROCS 1
+      STANDARD_PASS_OUTPUT
+      )
 
-  TRIBITS_ADD_EXECUTABLE_AND_TEST(
-    FadCommTests
-    SOURCES Fad_CommTests.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
-    ARGS
-    COMM serial mpi
-    STANDARD_PASS_OUTPUT
-  )
+    TRIBITS_ADD_EXECUTABLE_AND_TEST(
+      TaySerializationTests
+      SOURCES Tay_SerializationTests.cpp ${UTILS_SOURCES}
+      ARGS
+      COMM serial mpi
+      NUM_MPI_PROCS 1
+      STANDARD_PASS_OUTPUT
+      XHOSTTYPE Windows Windows_NT
+      )
 
-  TRIBITS_ADD_EXECUTABLE_AND_TEST(
-    ELRFadCommTests
-    SOURCES ELRFad_CommTests.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
-    ARGS
-    COMM serial mpi
-    STANDARD_PASS_OUTPUT
-  )
+    TRIBITS_ADD_EXECUTABLE_AND_TEST(
+      FadCommTests
+      SOURCES Fad_CommTests.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
+      ARGS
+      COMM serial mpi
+      STANDARD_PASS_OUTPUT
+      )
 
-  TRIBITS_ADD_EXECUTABLE_AND_TEST(
-    CacheFadCommTests
-    SOURCES CacheFad_CommTests.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
-    ARGS
-    COMM serial mpi
-    STANDARD_PASS_OUTPUT
-  )
+    TRIBITS_ADD_EXECUTABLE_AND_TEST(
+      ELRFadCommTests
+      SOURCES ELRFad_CommTests.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
+      ARGS
+      COMM serial mpi
+      STANDARD_PASS_OUTPUT
+      )
 
-  TRIBITS_ADD_EXECUTABLE_AND_TEST(
-    ELRCacheFadCommTests
-    SOURCES ELRCacheFad_CommTests.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
-    ARGS
-    COMM serial mpi
-    STANDARD_PASS_OUTPUT
-  )
+    TRIBITS_ADD_EXECUTABLE_AND_TEST(
+      CacheFadCommTests
+      SOURCES CacheFad_CommTests.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
+      ARGS
+      COMM serial mpi
+      STANDARD_PASS_OUTPUT
+      )
 
-  TRIBITS_ADD_EXECUTABLE_AND_TEST(
-    TayCommTests
-    SOURCES Tay_CommTests.cpp Tay_CommTests.hpp ${UTILS_SOURCES}
-    ARGS
-    COMM serial mpi
-    STANDARD_PASS_OUTPUT
-    XHOSTTYPE Windows Windows_NT
-  )
+    TRIBITS_ADD_EXECUTABLE_AND_TEST(
+      ELRCacheFadCommTests
+      SOURCES ELRCacheFad_CommTests.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
+      ARGS
+      COMM serial mpi
+      STANDARD_PASS_OUTPUT
+      )
+
+    TRIBITS_ADD_EXECUTABLE_AND_TEST(
+      TayCommTests
+      SOURCES Tay_CommTests.cpp Tay_CommTests.hpp ${UTILS_SOURCES}
+      ARGS
+      COMM serial mpi
+      STANDARD_PASS_OUTPUT
+      XHOSTTYPE Windows Windows_NT
+      )
+
+  ENDIF()
 
   IF (Sacado_ENABLE_KokkosCore)
 
     IF (Kokkos_ENABLE_Serial)
-      TRIBITS_ADD_EXECUTABLE_AND_TEST(
-        Fad_Kokkos_CommTests_Serial
-        SOURCES Fad_Kokkos_CommTests_Serial.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
-        ARGS
-        COMM serial mpi
-        STANDARD_PASS_OUTPUT
-        )
+      IF (Sacado_ENABLE_TeuchosKokkosComm)
+        TRIBITS_ADD_EXECUTABLE_AND_TEST(
+          Fad_Kokkos_CommTests_Serial
+          SOURCES Fad_Kokkos_CommTests_Serial.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
+          ARGS
+          COMM serial mpi
+          STANDARD_PASS_OUTPUT
+          )
+      ENDIF()
 
       TRIBITS_ADD_EXECUTABLE_AND_TEST(
         FadKokkosTests_Serial
@@ -108,13 +114,15 @@ IF (Sacado_ENABLE_Teuchos)
     ENDIF ()
 
     IF (Kokkos_ENABLE_Pthread)
-      TRIBITS_ADD_EXECUTABLE_AND_TEST(
-        Fad_Kokkos_CommTests_Threads
-        SOURCES Fad_Kokkos_CommTests_Threads.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
-        ARGS
-        COMM serial mpi
-        STANDARD_PASS_OUTPUT
-        )
+      IF (Sacado_ENABLE_TeuchosKokkosComm)
+        TRIBITS_ADD_EXECUTABLE_AND_TEST(
+          Fad_Kokkos_CommTests_Threads
+          SOURCES Fad_Kokkos_CommTests_Threads.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
+          ARGS
+          COMM serial mpi
+          STANDARD_PASS_OUTPUT
+          )
+      ENDIF()
 
       TRIBITS_ADD_EXECUTABLE_AND_TEST(
         FadKokkosTests_Threads
@@ -148,13 +156,15 @@ IF (Sacado_ENABLE_Teuchos)
     ENDIF()
 
     IF (Kokkos_ENABLE_OpenMP)
-      TRIBITS_ADD_EXECUTABLE_AND_TEST(
-        Fad_Kokkos_CommTests_OpenMP
-        SOURCES Fad_Kokkos_CommTests_OpenMP.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
-        ARGS
-        COMM serial mpi
-        STANDARD_PASS_OUTPUT
-        )
+      IF (Sacado_ENABLE_TeuchosKokkosComm)
+        TRIBITS_ADD_EXECUTABLE_AND_TEST(
+          Fad_Kokkos_CommTests_OpenMP
+          SOURCES Fad_Kokkos_CommTests_OpenMP.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
+          ARGS
+          COMM serial mpi
+          STANDARD_PASS_OUTPUT
+          )
+      ENDIF()
 
       TRIBITS_ADD_EXECUTABLE_AND_TEST(
         FadKokkosTests_OpenMP
@@ -188,13 +198,15 @@ IF (Sacado_ENABLE_Teuchos)
     ENDIF()
 
     IF (Kokkos_ENABLE_Cuda)
-      TRIBITS_ADD_EXECUTABLE_AND_TEST(
-        Fad_Kokkos_CommTests_Cuda
-        SOURCES Fad_Kokkos_CommTests_Cuda.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
-        ARGS
-        COMM serial mpi
-        STANDARD_PASS_OUTPUT
-        )
+      IF (Sacado_ENABLE_TeuchosKokkosComm)
+        TRIBITS_ADD_EXECUTABLE_AND_TEST(
+          Fad_Kokkos_CommTests_Cuda
+          SOURCES Fad_Kokkos_CommTests_Cuda.cpp Fad_CommTests.hpp ${UTILS_SOURCES}
+          ARGS
+          COMM serial mpi
+          STANDARD_PASS_OUTPUT
+          )
+      ENDIF()
 
       TRIBITS_ADD_EXECUTABLE_AND_TEST(
         FadKokkosTests_Cuda
