+++ packages/sacado/test/performance/fenl_assembly/CMakeLists.txt
@@ -4,7 +4,7 @@ INCLUDE_DIRECTORIES(REQUIRED_DURING_INSTALLATION_TESTING ${CMAKE_CURRENT_SOURCE_
 INCLUDE_DIRECTORIES(${CMAKE_CURRENT_BINARY_DIR})
 
 IF (Sacado_ENABLE_KokkosCore AND Sacado_ENABLE_KokkosContainers AND
-    Sacado_ENABLE_Teuchos)
+    Sacado_ENABLE_TeuchosCore)
 
   TRIBITS_ADD_EXECUTABLE(
     FadAssemblyTest
